requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
matplotlib>=3.2.0
pillow>=7.1.2
numpy>=1.18.5
